Market and Inventory are fully Func
Buying and selling items respectively adds or removes it from the farm
inventory and player bankroll is accordingly impact the player's
banckroll. Harvesting and planting also affect the farm inventory now.
As part of this mass change, inventory became part of the farm. it is
not longer a player field.